requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
email-validator
httpx
pytest
pytest-asyncio
